+++ apps/masterbots.ai/app/auth/signing/page.tsx
@@ -1,36 +1,8 @@
-'use client'
-
-import { useState } from 'react'
-import { signIn } from 'next-auth/react'
-import { useRouter } from 'next/navigation'
-import { LoginButton } from '@/components/login-button'
-import { Label } from '@/components/ui/label'
-import { Input } from '@/components/ui/input'
-import { Button } from '@/components/ui/button'
 import Link from 'next/link'
 import Image from 'next/image'
+import SignInForm from '@/components/login/signin-form'
 
-export default function SignIn() {
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-  const router = useRouter()
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    const result = await signIn('credentials', {
-      email,
-      password,
-      redirect: false
-    })
-
-    if (result?.error) {
-      // Handle error (show message to user)
-      console.error(result.error)
-    } else {
-      router.push('/dashboard') // Redirect to dashboard on success
-    }
-  }
-
+export default function SignInPage() {
   return (
     <div className="flex flex-col h-[calc(100vh-theme(spacing.16))] items-center py-10">
       <div
@@ -51,33 +23,7 @@ export default function SignIn() {
             Enter your email and password to access your account.
           </p>
         </div>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <Label htmlFor="email">Email</Label>
-            <Input
-              id="email"
-              type="email"
-              value={email}
-              onChange={e => setEmail(e.target.value)}
-              placeholder="m@example.com"
-              required
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="password">Password</Label>
-            <Input
-              id="password"
-              type="password"
-              value={password}
-              onChange={e => setPassword(e.target.value)}
-              required
-            />
-          </div>
-          <Button type="submit" className="w-full">
-            Sign In
-          </Button>
-          <LoginButton className="w-full" />
-        </form>
+        <SignInForm />
         <div className="space-y-4">
           <div className="text-sm text-center text-muted-foreground">
             Don&apos;t have an account?{' '}
